# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Socious-API
  version: 2.2.6
servers:
  - url: http://localhost:4000
    description: Local server (for development purposes)
  - url: https://devsif.socious.io
    description: dev server (for development purposes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
  /auth:
    post:
      summary: create auth
      description: create auth session
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              properties:
                redirect_url:
                  type: string
                  example: https://dev-id.socious.io
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
  /auth/session:
    post:
      summary: confirm session
      description: create auth session
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              properties:
                code:
                  type: string
                  example: 1234444
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
  /auth/refresh:
    post:
      summary: refresh jwt
      description: refresh jwt
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refresh_token:
                  type: string
                  example: jwt_refresh_token
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
...